+++ Stack/Opc.Ua.Core/Stack/Bindings/IFrameCaptureSink.cs
