+++ Stack/Opc.Ua.Diagnostics.Pcap/Models/CaptureSessionState.cs
@@ -27,6 +27,41 @@
  * http://opcfoundation.org/License/MIT/1.00/
  * ======================================================================*/
 
-using System;
+namespace Opc.Ua.Diagnostics.Pcap.Models
+{
+    /// <summary>
+    /// State a <c>CaptureSession</c> can be in.
+    /// </summary>
+    public enum CaptureSessionState
+    {
+        /// <summary>
+        /// The session has been created but not yet started.
+        /// </summary>
+        Starting = 0,
 
-[assembly: CLSCompliant(false)]
+        /// <summary>
+        /// The session is actively capturing.
+        /// </summary>
+        Running = 1,
+
+        /// <summary>
+        /// The session is in the middle of being stopped.
+        /// </summary>
+        Stopping = 2,
+
+        /// <summary>
+        /// The session stopped cleanly.
+        /// </summary>
+        Completed = 3,
+
+        /// <summary>
+        /// The session terminated because of an error.
+        /// </summary>
+        Failed = 4,
+
+        /// <summary>
+        /// The session has been disposed.
+        /// </summary>
+        Disposed = 5
+    }
+}
