+++ Stack/Opc.Ua.Diagnostics.Pcap/Models/FormatKind.cs
@@ -27,6 +27,41 @@
  * http://opcfoundation.org/License/MIT/1.00/
  * ======================================================================*/
 
-using System;
+namespace Opc.Ua.Diagnostics.Pcap.Models
+{
+    /// <summary>
+    /// Output formats supported by the diagnostics pipeline.
+    /// </summary>
+    public enum FormatKind
+    {
+        /// <summary>
+        /// libpcap (binary) - openable in Wireshark.
+        /// </summary>
+        Pcap = 0,
 
-[assembly: CLSCompliant(false)]
+        /// <summary>
+        /// PcapNg (binary) - openable in Wireshark.
+        /// </summary>
+        PcapNg = 1,
+
+        /// <summary>
+        /// Per-frame JSON.
+        /// </summary>
+        Json = 2,
+
+        /// <summary>
+        /// Per-frame CSV.
+        /// </summary>
+        Csv = 3,
+
+        /// <summary>
+        /// Human-readable per-frame text.
+        /// </summary>
+        Text = 4,
+
+        /// <summary>
+        /// OPC UA-specific decoded service-call timeline (text).
+        /// </summary>
+        ServiceTimeline = 5
+    }
+}
