+++ Stack/Opc.Ua.Diagnostics.Pcap/Models/NetworkInterfaceInfo.cs
@@ -27,6 +27,45 @@
  * http://opcfoundation.org/License/MIT/1.00/
  * ======================================================================*/
 
-using System;
+using System.Collections.Generic;
 
-[assembly: CLSCompliant(false)]
+namespace Opc.Ua.Diagnostics.Pcap.Models
+{
+    /// <summary>
+    /// Describes a local network interface returned by
+    /// <c>list_interfaces</c>.
+    /// </summary>
+    public sealed class NetworkInterfaceInfo
+    {
+        /// <summary>
+        /// Device name as understood by libpcap (e.g.
+        /// <c>\Device\NPF_{guid}</c>).
+        /// </summary>
+        public string Name { get; init; } = string.Empty;
+
+        /// <summary>
+        /// Friendly name (e.g. <c>Ethernet</c>).
+        /// </summary>
+        public string? FriendlyName { get; init; }
+
+        /// <summary>
+        /// Free-form description.
+        /// </summary>
+        public string? Description { get; init; }
+
+        /// <summary>
+        /// Bound IP addresses.
+        /// </summary>
+        public IReadOnlyList<string> Addresses { get; init; } = [];
+
+        /// <summary>
+        /// The link-layer type (e.g. <c>Ethernet</c>).
+        /// </summary>
+        public string? LinkType { get; init; }
+
+        /// <summary>
+        /// Whether this is the loopback interface.
+        /// </summary>
+        public bool IsLoopback { get; init; }
+    }
+}
