+++ Stack/Opc.Ua.Diagnostics.Pcap/Models/StartReplayRequest.cs
@@ -27,6 +27,51 @@
  * http://opcfoundation.org/License/MIT/1.00/
  * ======================================================================*/
 
-using System;
+using Opc.Ua.Diagnostics.Pcap.Replay;
 
-[assembly: CLSCompliant(false)]
+namespace Opc.Ua.Diagnostics.Pcap.Models
+{
+    /// <summary>
+    /// Parameters for replaying a pcap recording.
+    /// </summary>
+    public sealed class StartReplayRequest
+    {
+        /// <summary>
+        /// Path to the pcap file to replay.
+        /// </summary>
+        public string PcapFilePath { get; init; } = string.Empty;
+
+        /// <summary>
+        /// Optional keylog file used by mock-client replay for offline
+        /// decoding.
+        /// </summary>
+        public string? KeyLogFilePath { get; init; }
+
+        /// <summary>
+        /// Replay mode.
+        /// </summary>
+        public ReplayMode Mode { get; init; } = ReplayMode.MockServer;
+
+        /// <summary>
+        /// Listen URI scheme for mock-server replay. Defaults to
+        /// <c>opc.tcp</c>.
+        /// </summary>
+        public string? ListenScheme { get; init; }
+
+        /// <summary>
+        /// Listen port for mock-server replay. Defaults to an ephemeral
+        /// port.
+        /// </summary>
+        public int? ListenPort { get; init; }
+
+        /// <summary>
+        /// Target endpoint URL for mock-client replay.
+        /// </summary>
+        public string? TargetEndpointUrl { get; init; }
+
+        /// <summary>
+        /// Replay speed multiplier.
+        /// </summary>
+        public double Speed { get; init; } = 1.0;
+    }
+}
